fixed failing test linting
